# api/swagger.yml
openapi: "3.0.0"
info:
  description: Surogate Hub API
  title: Surogate Hub API
  license:
    name: "Apache 2.0"
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: "/api/v1"
    description: Surogate Hub server endpoint
security:
  - jwt_token: []
  - basic_auth: []
  - cookie_auth: []
  - oidc_auth: []
  - saml_auth: []
components:
  securitySchemes:
    basic_auth:
      type: http
      scheme: basic
    jwt_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookie_auth:
      type: apiKey
      in: cookie
      name: internal_auth_session
    oidc_auth:
      type: apiKey
      in: cookie
      name: oidc_auth_session
    saml_auth:
      type: apiKey
      in: cookie
      name: saml_auth_session
  parameters:
    PaginationPrefix:
      in: query
      name: prefix
      description: return items prefixed with this value
      schema:
        type: string
    PaginationAfter:
      in: query
      name: after
      description: return items after this value
      schema:
        type: string
    PaginationAmount:
      in: query
      name: amount
      description: how many items to return
      schema:
        type: integer
        minimum: -1
        maximum: 1000
        default: 100
    PaginationDelimiter:
      in: query
      name: delimiter
      description: delimiter used to group common prefixes by
      schema:
        type: string
...
    IfNoneMatch:
      in: header
      name: If-None-Match
      description: Set to "*" to atomically allow the upload only if the key has no object yet. Other values are not supported.
      example: "*"
      required: false
      schema:
        type: string
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    PreconditionFailed:
      description: Precondition Failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Pagination:
      type: object
      required:
        - has_more
        - max_per_page
        - results
        - next_offset
      properties:
        has_more:
          type: boolean
          description: Next page is available
        next_offset:
          type: string
          description: Token used to retrieve the next page
        results:
          type: integer
          minimum: 0
          description: Number of values found in the results
        max_per_page:
          type: integer
          minimum: 0
          description: Maximal number of entries per page
...
    PathList:
      type: object
      required:
        - paths
      properties:
        paths:
          type: array
          items:
            type: string
            description: Object path
    ObjectStats:
      type: object
      required:
        - checksum
        - physical_address
        - path
        - path_type
        - mtime
      properties:
        path:
          type: string
        path_type:
          type: string
          enum: [common_prefix, object]
        physical_address:
          type: string
          description: |
            The location of the object on the underlying object store.
            Formatted as a native URI with the object store type as scheme ("s3://...", "gs://...", etc.)
            Or, in the case of presign=true, will be an HTTP URL to be consumed via regular HTTP GET
        physical_address_expiry:
          type: integer
          format: int64
          description: |
            If present and nonzero, physical_address is a pre-signed URL and
            will expire at this Unix Epoch time.  This will be shorter than
            the pre-signed URL lifetime if an authentication token is about
            to expire.

            This field is *optional*.
        checksum:
          type: string
        size_bytes:
          type: integer
          format: int64
          description: |
            The number of bytes in the object.  Surogate Hub always populates this
            field when returning ObjectStats.  This field is optional _for
            the client_ to supply, for instance on upload.
        mtime:
          type: integer
          format: int64
          description: Unix Epoch in seconds
        metadata:
          $ref: "#/components/schemas/ObjectUserMetadata"
        content_type:
          type: string
          description: Object media type
    ObjectStatsList:
      type: object
      required:
        - pagination
        - results
      properties:
        pagination:
          $ref: "#/components/schemas/Pagination"
        results:
          type: array
          items:
            $ref: "#/components/schemas/ObjectStats"
...
    ObjectStageCreation:
      type: object
      required:
        - physical_address
        - checksum
        - size_bytes
      properties:
        physical_address:
          type: string
        checksum:
          type: string
        size_bytes:
          type: integer
          format: int64
        mtime:
          type: integer
          format: int64
          description: Unix Epoch in seconds
        metadata:
          $ref: "#/components/schemas/ObjectUserMetadata"
        content_type:
          type: string
          description: Object media type
        force:
          type: boolean
          default: false
    ObjectUserMetadata:
      type: object
      additionalProperties:
        type: string
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          description: short message explaining the error
          type: string
    ObjectError:
      type: object
      required:
        - status_code
        - message
      properties:
        status_code:
          type: integer
          description: HTTP status code associated for operation on path
        message:
          type: string
          description: short message explaining status_code
        path:
          type: string
          description: affected path
    ObjectErrorList:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ObjectError"
...
    User:
      type: object
      required:
        - id
        - creation_date
      properties:
        id:
          type: string
          description: A unique identifier for the user. Cannot be edited.
        creation_date:
          type: integer
          format: int64
          description: Unix Epoch in seconds
        friendly_name:
          type: string
          description: |
            A shorter name for the user than the id. Unlike id it does not identify the user (it
            might not be unique). Used in some places in the UI.
        email:
          type: string
          description: |
            The email address of the user. If API authentication is enabled, this field is mandatory and will be invited to login.
            If API authentication is disabled, this field will be ignored. All current APIAuthenticators require the email to be 
            lowercase and unique, although custom authenticators may not enforce this.
...
paths:
...
  /auth/users/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: string
    get:
      tags:
        - auth
      operationId: getUser
      summary: get user
      responses:
        200:
          description: user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        420:
          description: too many requests
        default:
          $ref: "#/components/responses/ServerError"
    delete:
      tags:
        - auth
      operationId: deleteUser
      summary: delete user
      responses:
        204:
          description: user deleted successfully
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        420:
          description: too many requests
        default:
          $ref: "#/components/responses/ServerError"
...
  /repositories/{repository}/refs/{ref}/objects:
    parameters:
      - in: path
        name: repository
        required: true
        schema:
          type: string
      - in: path
        name: ref
        required: true
        schema:
          type: string
        description: a reference (could be either a branch or a commit ID)
      - in: query
        name: path
        description: relative to the ref
        required: true
        schema:
          type: string
    get:
      tags:
        - objects
      operationId: getObject
      summary: get object content
      parameters:
        - in: header
          name: Range
          description: Byte range to retrieve
          example: "bytes=0-1023"
          required: false
          schema:
            type: string
            pattern: '^bytes=((\d*-\d*,? ?)+)$'
        - in: header
          name: If-None-Match
          description: Returns response only if the object does not have a matching ETag
          example: "33a64df551425fcc55e4d42a148795d9f25f89d4"
          required: false
          schema:
            type: string
        - in: query
          name: presign
          required: false
          schema:
            type: boolean
      responses:
        200:
          description: object content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
            Last-Modified:
              schema:
                type: string
            ETag:
              schema:
                type: string
        206:
          description: partial object content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
            Content-Range:
              schema:
                type: string
                pattern: '^bytes=((\d*-\d*,? ?)+)$'
            Last-Modified:
              schema:
                type: string
            ETag:
              schema:
                type: string
        302:
          description: Redirect to a pre-signed URL for the object
          headers:
            Location:
              schema:
                type: string
        304:
          description: Content not modified
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        410:
          description: object expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        416:
          description: Requested Range Not Satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        420:
          description: too many requests
        default:
          $ref: "#/components/responses/ServerError"
    head:
      tags:
        - objects
      operationId: headObject
      summary: check if object exists
      parameters:
        - in: header
          name: Range
          description: Byte range to retrieve
          example: "bytes=0-1023"
          required: false
          schema:
            type: string
            pattern: '^bytes=((\d*-\d*,? ?)+)$'
      responses:
        200:
          description: object exists
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
            Last-Modified:
              schema:
                type: string
            ETag:
              schema:
                type: string
        206:
          description: partial object content info
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
            Content-Range:
              schema:
                type: string
                pattern: '^bytes=((\d*-\d*,? ?)+)$'
            Last-Modified:
              schema:
                type: string
            ETag:
              schema:
                type: string
        401:
          description: Unauthorized
        404:
          description: object not found
        410:
          description: object expired
        416:
          description: Requested Range Not Satisfiable
        420:
          description: too many requests
        default:
          description: internal server error
...
  /repositories/{repository}/branches/{branch}/objects:
    parameters:
      - in: path
        name: repository
        required: true
        schema:
          type: string
      - in: path
        name: branch
        required: true
        schema:
          type: string
      - in: query
        name: path
        description: relative to the branch
        required: true
        schema:
          type: string
    put:
      deprecated: true
      tags:
        - internal
      operationId: stageObject
      summary: stage an object's metadata for the given branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjectStageCreation"
      responses:
        201:
          description: object metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectStats"
        400:
          $ref: "#/components/responses/ValidationError"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        420:
          description: too many requests
        default:
          $ref: "#/components/responses/ServerError"

    post:
      tags:
        - objects
      operationId: uploadObject
      x-validation-exclude-body: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                content:
                  description: Only a single file per upload which must be named "content".
                  type: string
                  format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary

      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
        - in: query
          name: storageClass
          description: Deprecated, this capability will not be supported in future releases.
          required: false
          deprecated: true
          schema:
            type: string
        - in: query
          name: force
          required: false
          schema:
            type: boolean
            default: false
      responses:
        201:
          description: object metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectStats"
        400:
          $ref: "#/components/responses/ValidationError"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        412:
          $ref: "#/components/responses/PreconditionFailed"
        420:
          description: too many requests
        default:
          $ref: "#/components/responses/ServerError"
    delete:
      tags:
        - objects
      operationId: deleteObject
      summary: delete object. Missing objects will not return a NotFound error.
      parameters:
        - in: query
          name: force
          required: false
          schema:
            type: boolean
            default: false
      responses:
        204:
          description: object deleted successfully
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        420:
          description: too many requests
        default:
          $ref: "#/components/responses/ServerError"
  /repositories/{repository}/branches/{branch}/objects/delete:
    parameters:
      - in: path
        name: repository
        required: true
        schema:
          type: string
      - in: path
        name: branch
        required: true
        schema:
          type: string
      - in: query
        name: force
        required: false
        schema:
          type: boolean
          default: false
    post:
      tags:
        - objects
      operationId: deleteObjects
      summary: delete objects. Missing objects will not return a NotFound error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PathList"
      responses:
        200:
          description: Delete objects response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectErrorList"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        420:
          description: too many requests
        default:
          $ref: "#/components/responses/ServerError"
...
  /repositories/{repository}/refs/{ref}/objects/ls:
    parameters:
      - in: path
        name: repository
        required: true
        schema:
          type: string
      - in: path
        name: ref
        required: true
        schema:
          type: string
        description: a reference (could be either a branch or a commit ID)
      - in: query
        name: user_metadata
        required: false
        schema:
          type: boolean
          default: true
      - in: query
        name: presign
        required: false
        schema:
          type: boolean
      - $ref: "#/components/parameters/PaginationAfter"
      - $ref: "#/components/parameters/PaginationAmount"
      - $ref: "#/components/parameters/PaginationDelimiter"
      - $ref: "#/components/parameters/PaginationPrefix"

    get:
      tags:
        - objects
      operationId: listObjects
      summary: list objects under a given prefix
      responses:
        200:
          description: object listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectStatsList"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        420:
          description: too many requests
        default:
          $ref: "#/components/responses/ServerError"
...